+++ src/main/java/com/mock/investment/stock/domain/order/dto/SellOrderDto.java
@@ -21,13 +21,13 @@ public class SellOrderDto {
     private Double buyPrice;
 
     @Schema(description = "주문 수량", example = "10")
-    private Integer buyQuantity;
+    private Double buyQuantity;
 
     @Schema(description = "매수 체결 수량", example = "3")
-    private Integer filledQuantity;
+    private Double filledQuantity;
 
     @Schema(description = "매수 주문 남은 수량", example = "7")
-    private Integer remainingQuantity;
+    private Double remainingQuantity;
 
     public static SellOrderDto fromEntity(SellOrder sellOrder) {
         return SellOrderDto.builder()
